ssh-agent: Add support to load additional certificates
Add support to load additional certificates
for already loaded privated keys. Useful
if the private key is on a PKCS#11 hardware token.

The private keys inside ssh-agent are now refcounted
to share the private parts between "Identities".
The reason for this change was that the PKCS#11 code
might have redirected ("wrap") the RSA functions to a hardware token.
We don't want to mess with those internals.